% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{find_df_pkgs}
\alias{find_df_pkgs}
\title{Find packages with data.frames}
\usage{
find_df_pkgs()
}
\value{
named vector of packages with data.frames
}
\description{
Return a named character vector of installed package with \code{data.frame}s.
}
\examples{
find_df_pkgs()
}
