% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{find_pkg_df_nms}
\alias{find_pkg_df_nms}
\title{Find names of data.frame in a package}
\usage{
find_pkg_df_nms(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
named vector of data.frames in package
}
\description{
Return the names of \code{data.frame}s in a package.
}
\examples{
find_pkg_df_nms(pkg = "base")
find_pkg_df_nms(pkg = "datasets")
find_pkg_df_nms(pkg = "lubridate")
}
