% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{get_pkg_df}
\alias{get_pkg_df}
\title{Get data.frame from package}
\usage{
get_pkg_df(df, pkg)
}
\arguments{
\item{df}{name of dataset}

\item{pkg}{name of package}
}
\value{
data object
}
\description{
Get data.frame from package
}
\examples{
get_pkg_df("gss_cat", "forcats")
get_pkg_df("starwars", "dplyr")
}
