% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistogramServer.R
\name{gg_hist}
\alias{gg_hist}
\title{Histogram plot (with ggplot2)}
\usage{
gg_hist(df, x_var, bins, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{bins}{from \code{ggplot2::geom_histogram()} ('Number of bins. Overridden by binwidth. Defaults to 30')}

\item{...}{other arguments passed to \code{ggplot2::geom_histogram()}, outside of \code{ggplot2::aes()}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Histogram plot (with ggplot2)
}
\examples{
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), 10000), ]
gg_hist(
  df = mini_dmnds,
  x_var = "carat",
  bins = 45,
  alpha = 1 / 3,
  color = "#772953"
) +
  ggplot2::labs(title =
      gg_x_dist_title(x = "carat"))
}
