% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistogramServer.R
\name{gg_x_dist_title}
\alias{gg_x_dist_title}
\title{Distribution of (x) plot title}
\usage{
gg_x_dist_title(x)
}
\arguments{
\item{x}{x variable}
}
\value{
String for plot title
}
\description{
Distribution of (x) plot title
}
\examples{
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), 10000), ]
gg_base(df = mini_dmnds, x_var = "carat") +
  ggplot2::geom_histogram(bins = 30) +
  ggplot2::labs(title =
      gg_x_dist_title(x = "carat"))
}
