% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistApp.R
\name{gghistApp}
\alias{gghistApp}
\title{ggplot2 Histogram App}
\usage{
gghistApp(pkg = NULL)
}
\arguments{
\item{pkg}{packages listing \code{data.frame}s}
}
\value{
shiny app
}
\description{
This module function has been adapted from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{Case study: histogram} section of Mastering Shiny.
}
\section{Histogram App}{

The \code{gghistogramApp()} is built using the \code{pkgDatasetInput()}/\code{pkgDatasetServer()},
\code{selectVarInput()}/\code{selectVarServer()} and \code{histogramOutput()}/\code{gghistServer()}
modules.
}

\examples{
gghistApp(pkg = c("dplyr", "palmerpenguins"))
}
