% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistServer.R
\name{gghistServer}
\alias{gghistServer}
\title{Histogram server module (ggplot2)}
\usage{
gghistServer(id, x, title = reactive("Histogram"))
}
\arguments{
\item{id}{shiny module server id}
}
\value{
shiny server module
}
\description{
This module function is an adapted version of \code{histogramServer()} from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{Case study: histogram} section of Mastering Shiny.
}
