% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msst2ap-package.R
\docType{package}
\name{msst2ap-package}
\alias{msst2ap}
\alias{msst2ap-package}
\title{msst2ap: Test Your Shiny Package App server functions}
\description{
This shiny app-package if from Mastering Shiny and was developed to demonstrate testing with testthat and shiny's testServer().
}
\keyword{internal}
