% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{pkg_df_check}
\alias{pkg_df_check}
\title{Check if package contains data.frames}
\usage{
pkg_df_check(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\value{
logical (\code{TRUE} = has data.frame, \code{FALSE} = no data.frame)
}
\description{
Returns \code{TRUE} if package has \code{data.frame}. If package is not installed,
install with \code{install.packages()}.
\enumerate{
\item Check if the package is installed and load it
\item Retrieve the objects in the package
\item Use \code{purrr::map_lgl()} to apply \code{is.data.frame()} to each object in the
package. \code{map_lgl()} returns a logical vector with the same length as
the retrieved package objects.
}
}
\examples{
pkg_df_check("dplyr")
pkg_df_check("stringr")
}
