% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{pkg_inst_check}
\alias{pkg_inst_check}
\title{Package installation check (base)}
\usage{
pkg_inst_check(pkg, quiet = TRUE)
}
\arguments{
\item{pkg}{package or packages}
}
\value{
installation success or failure
}
\description{
Check if a package is installed, if not, install it.
}
\examples{
p <- c("emoji", "haven")
 if (any(p  \%in\% loadedNamespaces())) {
  remove.packages(p)
 }
pkg_inst_check(c("emoji", "haven"))
}
