% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_dfs.R
\name{pkg_install_check}
\alias{pkg_install_check}
\title{Check if a package is installed, if not install it}
\usage{
pkg_install_check(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\description{
Check if \code{pkg} is installed with \code{rlang::is_installed()}. If not, package
is installed with \code{pak::pkg_install()}.
}
\examples{
p <- c("emoji", "palmerpenguins")
 if (any(p  \%in\% loadedNamespaces())) {
  remove.packages(p)
 }
pkg_install_check(c("janitor", "snakecase"))
}
