% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_ns_check}
\alias{pkg_ns_check}
\title{Check if package namespace is loaded, if not load it}
\usage{
pkg_ns_check(pkg, quiet = FALSE)
}
\arguments{
\item{pkg}{name of package (a character vector)}

\item{quiet}{logical, print messages}
}
\value{
Package: \code{'name'} loaded or Loading package: \code{'name'}
}
\description{
Check if \code{pkg} is installed with \code{isNamespaceLoaded()}. If not, package is
installed with \code{requireNamespace(quietly = TRUE)}.
}
\examples{
# remove from ns
unloadNamespace("fs")
unloadNamespace("box")
# with single pkg
pkg_ns_check("fs")
# remove again
unloadNamespace("fs")
# with multiple pkgs
pkgs <- c("fs", "box")
pkg_ns_check(pkgs)
}
