% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgData.R
\name{pkgDataInput}
\alias{pkgDataInput}
\title{Package data input module (UI)}
\usage{
pkgDataInput(id, pkg = NULL)
}
\arguments{
\item{id}{module id}

\item{pkg}{packages to install}
}
\value{
UI module function
}
\description{
This module includes a package argument for installing and loading additional
data objects. Its been adapted from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output}
section of Mastering Shiny.
}
