% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgData.R
\name{pkgDataServer}
\alias{pkgDataServer}
\title{Package data module (server)}
\usage{
pkgDataServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
module server function (for package datasets).
}
\description{
Package data module (server)
}
