% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDatasetApp.R
\name{pkgdatasetApp}
\alias{pkgdatasetApp}
\title{Package Dataset App}
\usage{
pkgdatasetApp(pkg = NULL)
}
\value{
shiny application
}
\description{
This standalone app function has been adapted from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
\section{Packages}{

This application includes an argument for entering the packages to display in
the \code{selectInput()} in the UI.
}

\section{reactive values}{

The \code{renderPrint()} output contains the output from \code{shiny::reactiveValuesToList()}
and contains the\code{inputId}s for the modules in the application.
}

