% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDataset.R
\name{pkgDatasetServer}
\alias{pkgDatasetServer}
\title{Package data module (server)}
\usage{
pkgDatasetServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
module server function (for package datasets).
}
\description{
Package data module (server)
}
