% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msst2ap-package.R
\docType{package}
\name{msst2ap-package}
\alias{msst2ap}
\alias{msst2ap-package}
\title{msst2ap: Testing Your Shiny App-Package}
\description{
This shiny app-package if from Mastering Shiny and was developed to demonstrate testing with testthat and shinytest2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/msst2ap/}
}

}
\keyword{internal}
